Seperate login dialog into component
